+++ src/app/components/core/control-dispatcher/list-control-dispatcher/acl-list/acl-list.component.ts
@@ -1,32 +1,32 @@
-import { Component, OnInit, ViewChild, EventEmitter } from '@angular/core';
-import { Router } from '@angular/router';
+import { Component, OnInit, ViewChild, EventEmitter, OnDestroy } from '@angular/core';
+import { Router, ActivatedRoute } from '@angular/router';
 import { Subscription } from 'rxjs';
+import { map as _map } from 'rxjs/operators';
 import { AlertService, SharedService, ProgressBarService, RolesService } from '../../../../../services';
 import { ConfirmationDialogComponent } from '../../../../common/confirmation-dialog/confirmation-dialog.component';
 import { DialogService } from '../../../../common/confirmation-dialog/dialog.service';
-import { orderBy } from 'lodash';
+import { orderBy, map } from 'lodash';
 import { AclService } from '../../../../../services/acl.service';
 import { DocService } from '../../../../../services/doc.service';
-import { AddDispatcherServiceComponent } from './../../add-dispatcher-service/add-dispatcher-service.component';
+import { AdditionalServicesUtils } from '../../../developer/additional-services/additional-services-utils.service';
 
 @Component({
   selector: 'app-acl-list',
   templateUrl: './acl-list.component.html',
   styleUrls: ['./acl-list.component.css']
 })
-export class AclListComponent implements OnInit {
+export class AclListComponent implements OnInit, OnDestroy {
   @ViewChild('confirmationDialog') confirmationDialog: ConfirmationDialogComponent;
-  @ViewChild(AddDispatcherServiceComponent, { static: true }) addDispatcherServiceComponent: AddDispatcherServiceComponent;
 
   controlAcls: any = [];
-  
+
   acl;
-  isServiceAvailable = false;
   private subscription: Subscription;
+  private serviceDetailsSubscription: Subscription;
+  private paramsSubscription: Subscription;
 
   public reenableButton = new EventEmitter<boolean>(false);
-  showConfigureModal = false;
-  serviceInfo = {};
+  serviceInfo = { added: false, isEnabled: true, name: '', isInstalled: false };
 
   constructor(
     private aclService: AclService,
@@ -36,13 +36,32 @@ export class AclListComponent implements OnInit {
     private router: Router,
     public rolesService: RolesService,
     public sharedService: SharedService,
+    public activatedRoute: ActivatedRoute,
+    private additionalServicesUtils: AdditionalServicesUtils,
     private ngProgress: ProgressBarService) {
+    // If we are redirecting back after enabling/disabling/adding the service then no need to make all calls again
+    this.paramsSubscription = this.activatedRoute.paramMap
+      .pipe(_map(() => window.history.state)).subscribe((data: any) => {
+        if (!data?.shouldSkipCalls) {
+          this.additionalServicesUtils.getAllServiceStatus(false, 'dispatcher');
+        }
+      })
+    // Issue may cause by refreshing the page because of old state data, so need to update history state
+    history.replaceState({ shouldSkipCalls: false }, '');
   }
 
   ngOnInit(): void {
+    this.serviceDetailsSubscription = this.sharedService.installedServicePkgs.subscribe(service => {
+      if (service.installed) {
+        this.serviceInfo = service.installed;
+      }
+    });
     this.getACLs();
   }
 
+  refreshServiceInfo() {
+    this.additionalServicesUtils.getAllServiceStatus(false, 'dispatcher');
+  }
 
   getACLs() {
     /** request started */
@@ -134,37 +153,23 @@ export class AclListComponent implements OnInit {
       });
   }
 
-  getServiceDetail(event) {
-    this.showConfigureModal = event.isOpen;
-    delete event.isOpen;
-    this.serviceInfo = event;
-    if (this.showConfigureModal) {
-      this.openServiceConfigureModal();
-    }
-  }
-
   /**
    * Open Configure Service modal
    */
   openServiceConfigureModal() {
-    this.router.navigate(['/developer/options/additional-services/config'], { state: { ...this.serviceInfo }});
+    this.router.navigate(['/developer/options/additional-services/config'], { state: { ...this.serviceInfo } });
   }
 
   goToLink(urlSlug: string) {
     this.docService.goToSetPointControlDocLink(urlSlug);
   }
 
-  onNotify(handleEvent) {
-    if (handleEvent) {
-      this.addDispatcherServiceComponent.getInstalledServicesList();
-    }
-    return;
-  }
-
   public ngOnDestroy(): void {
     if (this.subscription) {
       this.subscription.unsubscribe();
     }
+    this.serviceDetailsSubscription.unsubscribe();
+    this.paramsSubscription.unsubscribe();
   }
 }