+++ src/app/components/core/notifications/notifications.module.ts
@@ -14,7 +14,6 @@ import { ServicesApiService, NotificationsService } from '../../../services';
 import { AddNotificationWizardComponent } from './add-notification-wizard/add-notification-wizard.component';
 import { NotificationModalComponent } from './notification-modal/notification-modal.component';
 import { DeveloperModule } from '../developer/developer.module';
-import { ServiceResolver } from '../../../resolver/service.resolver';
 import { NumberInputDebounceModule } from '../../common/number-input-debounce/number-input-debounce.module';
 import { PaginationModule } from '../../common/pagination/pagination.module';
 import { FlowEditorModule } from '../../common/node-editor/flow-editor.module';
@@ -22,8 +21,7 @@ import { FlowEditorModule } from '../../common/node-editor/flow-editor.module';
 const routes: Routes = [
   {
     path: '',
-    component: NotificationsComponent,
-    resolve: { service: ServiceResolver }
+    component: NotificationsComponent
   },
   {
     path: 'add',