+++ docs/changelog.md
@@ -1,3 +1,31 @@
+## v2.6.0 [2024-10-24]
+
+### Added
+- [FOGL-8841]: Manage Configuration items access for users, per optionally available role based permissions list [#437](https://github.com/fledge-iot/fledge-gui/pull/437)
+
+### Changed
+- [FOGL-8942]: Configuration items group tabs are sorted by the minimum order of config item in the group [#471](https://github.com/fledge-iot/fledge-gui/pull/471)
+- [FOGL-8458]: Developer page access through developer sub-menu options [#456](https://github.com/fledge-iot/fledge-gui/pull/456)
+- [FOGL-8990]: Enhanced support for list/kvlist type configuration item [#462](https://github.com/fledge-iot/fledge-gui/pull/462)
+  - Add & Delete 
+  - Expand/Collapse individual item
+  - Expand All/Collapse All
+
+### Fixed
+- [FOGL-8588]: Filter plugin installation issue where plugin name and package name had a mismatch [#467](https://github.com/fledge-iot/fledge-gui/pull/467)
+- [FOGL-9080]: Missing child categories issue in category dropdown on add control script page [#460](https://github.com/fledge-iot/fledge-gui/pull/460)
+- [FOGL-9083]: Control API entry point access issue for less privileged user [#458](https://github.com/fledge-iot/fledge-gui/pull/458)
+- [FOGL-9068]: Allow retrigger time of 0 for notification [#455](https://github.com/fledge-iot/fledge-gui/pull/455)
+- [FOGL-8803]: Deletion of persisted data in Developer mode [#442](https://github.com/fledge-iot/fledge-gui/pull/442)
+- [FOGL-8891]: Log out active session issue on user profile page [#441](https://github.com/fledge-iot/fledge-gui/pull/441)
+
+### Others
+- Visual programming based Flow Editor interface improvements
+  - Delete option on connection line between nodes
+  - Undo/Redo functionality
+  - Filter drag & drop on connection line directly 
+- UI/UX improvement and various functional bug fixes
+
 ## v2.5.0 [2024-06-26]
 
 ### Added
@@ -17,7 +45,6 @@
 ### Others
 - New & updated Brand logo & favicon [#403](https://github.com/fledge-iot/fledge-gui/pull/403)
 - UI/UX improvement and various functional bug fixes.
-
 
 
 ## v2.4.0 [2024-04-17]